test: split attendees from optional attendees tests
