+++ src/view/editor/parts/tests/editor-attendees.test.tsx
@@ -350,158 +350,4 @@ describe('Editor Attendees', () => {
 			});
 		});
 	});
-
-	describe('Optional Attendees', () => {
-		describe('ChipInput', () => {
-			it('should display optional attendees', async () => {
-				const store = configureStore({ reducer: combineReducers(reducers) });
-				const editor = generateEditor({
-					context: {
-						dispatch: store.dispatch,
-						folders: {},
-						optionalAttendees: [
-							{ email: 'email1@test.com', label: 'Optional Test 1' },
-							{ email: 'email1@test.com', label: 'Optional Test 2' }
-						]
-					}
-				});
-
-				setupTest(<EditorAttendees editorId={editor.id} />, { store });
-
-				expect(screen.getByText('Optional Test 2')).toBeVisible();
-				expect(screen.getByText('Optional Test 1')).toBeVisible();
-			});
-
-			it('should not clear existing optional attendees after adding a new one', async () => {
-				const store = configureStore({ reducer: combineReducers(reducers) });
-
-				const editor = generateEditor({
-					context: {
-						dispatch: store.dispatch,
-						folders: {},
-						optionalAttendees: [
-							{ email: 'email1@test.com', label: 'Test 1' },
-							{ email: 'email2@test.com', label: 'Test 2' }
-						]
-					}
-				});
-				const { user } = setupTest(<EditorAttendees editorId={editor.id} />, { store });
-				const chipInput = await screen.findByTestId('optional-attendees-chip-input');
-				const attendees = await screen.findAllByText('Attendees');
-
-				await user.type(within(chipInput).getByRole('textbox'), 'email3@test.com');
-				await user.click(attendees[0]);
-
-				expect(await screen.findByText('email3@test.com')).toBeInTheDocument();
-				expect(screen.getByText('Test 2')).toBeVisible();
-				expect(screen.getByText('Test 1')).toBeVisible();
-			});
-
-			it('should display optional attendee label in chip if label available', () => {
-				const store = configureStore({ reducer: combineReducers(reducers) });
-				const editor = generateEditor({
-					context: {
-						dispatch: store.dispatch,
-						folders: {},
-						optionalAttendees: [{ email: 'email1@test.com', label: 'Test 1' }]
-					}
-				});
-
-				setupTest(<EditorAttendees editorId={editor.id} />, { store });
-
-				expect(screen.getByText('Test 1')).toBeVisible();
-			});
-
-			it('should display optional attendee email in chip if label not available', () => {
-				const store = configureStore({ reducer: combineReducers(reducers) });
-				const editor = generateEditor({
-					context: {
-						dispatch: store.dispatch,
-						folders: {},
-						optionalAttendees: [{ email: 'email1@test.com' }]
-					}
-				});
-				setupTest(<EditorAttendees editorId={editor.id} />, { store });
-
-				expect(screen.getByText('email1@test.com')).toBeVisible();
-			});
-		});
-
-		describe('ContactInput', () => {
-			it('should display edit action when new value in ContactInput has an error', async () => {
-				const store = configureStore({ reducer: combineReducers(reducers) });
-				jest.spyOn(shellUi, 'useIntegratedComponent').mockReturnValue([ContactInputError, true]);
-				const editor = generateEditor({
-					context: {
-						dispatch: store.dispatch,
-						folders: {},
-						optionalAttendees: [{ email: 'email1@test.com', label: 'Test 1' }]
-					}
-				});
-				const { user } = setupTest(<EditorAttendees editorId={editor.id} />, { store });
-				const testButton = await screen.findAllByTestId('test-button');
-				await user.click(testButton[1]);
-
-				await waitFor(() => {
-					expect(spyDefaultValue).toBeCalledWith(
-						expect.arrayContaining([
-							expect.objectContaining({
-								error: true,
-								actions: [editAction]
-							})
-						])
-					);
-				});
-			});
-
-			it('should not display edit action when new value in ContactInput has no errors', async () => {
-				const store = configureStore({ reducer: combineReducers(reducers) });
-				jest.spyOn(shellUi, 'useIntegratedComponent').mockReturnValue([ContactInput, true]);
-				const editor = generateEditor({
-					context: {
-						dispatch: store.dispatch,
-						folders: {},
-						optionalAttendees: [{ email: 'email1@test.com', label: 'Test 1' }]
-					}
-				});
-				const { user } = setupTest(<EditorAttendees editorId={editor.id} />, { store });
-				const testButton = await screen.findAllByTestId('test-button');
-				await user.click(testButton[1]);
-
-				await waitFor(() => {
-					expect(spyDefaultValue).toBeCalledWith(
-						expect.arrayContaining([
-							expect.objectContaining({
-								error: false,
-								actions: []
-							})
-						])
-					);
-				});
-			});
-
-			it('should not pass firstName, lastName, label to the ContactInput component', async () => {
-				jest.spyOn(shellUi, 'useIntegratedComponent').mockReturnValue([ContactInput, true]);
-				const store = configureStore({ reducer: combineReducers(reducers) });
-				const editor = generateEditor({
-					context: {
-						dispatch: store.dispatch,
-						folders: {},
-						optionalAttendees: [{ email: 'email1@test.com', fullName: 'Test 1' }]
-					}
-				});
-				setupTest(<EditorAttendees editorId={editor.id} />, { store });
-
-				expect(spyDefaultValue).toHaveBeenCalledWith([
-					{
-						fullName: 'Test 1',
-						email: 'email1@test.com',
-						actions: undefined,
-						error: undefined,
-						id: undefined
-					}
-				]);
-			});
-		});
-	});
 });
